// types/IDetailedPostcodesItem.ts
// generated typescript definitions from database using groovy script


export interface IDetailedPostcodesItem {

   postcode: string;
   in_use: string;
   lat: number;
   long: number;
   grid_ref: string;
   county: string;
   district: string;
   ward: string;
   district_code: string;
   ward_code: string;
   country: string;
   county_code: string;
   constituency: string;
   introduced: Date;
   terminated: Date;
   parish: string;
   national_park: string;
   population: number;
   households: number;
   built_up_area: string;
   built_up_sub_division: string;
   lower_layer_super_output_area: string;
   rural_urban: string;
   region: string;
   Altitude: number;
   local_authority: string;
   msoa_code: string;
   msoa: string;
   parish_code: string;
   census_output_area: string;
   constituency_code: string;
   lat_updated: Date;
   nearest_station: string;
   postcode_prefix: string;
   postcode_district: string;
   plus_code: string;
   average_income: string;

}